Refactored google sign in and google sign out into a new component directory
